Update source-code.py
Added some locations where you may want to add more documentation for someone who would download this and try to run it or change the code for their own experimental purposes. 

Three additional things as we go to push this online, too. The first is saving the data. Right now it saves the file locally. But, we can simply modify the code to send the data to an online storage service like Amazon S3, Google Cloud Storage, or Azure Blob Storage via the respective SDKs (like boto3 for AWS S3). 

The second is the display of the participants data. Up to you if you want to make this a graph the user sees at the end. Might be kind of neat to show them that. 

The last bit is we need to create a landing page for the experiment. The page should be where we have a brief description of the experiment along with the informed consent for participating. Here is a link to the informed consent we would modify specific to this experiment: https://endicott.qualtrics.com/jfe/form/SV_01Zn1I0kr9qoOMK

Let me know if you have any questions as we get moving!